# frontend/src/api/openapi.json
{
  "openapi": "3.0.0",
  "components": {
...
    "schemas": {
      "SuccessResponseType": {
        "description": "Wrapper de éxito JSend (aplicado por el middleware global).",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success"
            ],
            "nullable": false
          },
          "data": {
            "properties": {},
            "type": "object"
          }
        },
        "required": [
          "status",
          "data"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
      "FailResponseFromError_EmailAlreadyInUseError_": {
        "$ref": "#/components/schemas/FailResponse_EMAIL_ALREADY_IN_USE._field-email--value-string__",
        "description": "Extrae automáticamente code y details de una clase AppError.\r\n\r\nEjemplo:\r\n- FailResponseFromError<UserAlreadyExistsError> → { status: 'fail', data: { code: 'CONFLICT', message: string, details: { field, value } } }\r\n- FailResponseFromError<UserNotFoundError> → { status: 'fail', data: { code: 'NOT_FOUND', message: string, details: { userId } } }"
      },
...
      "InitiateRegistrationRequestValidationFailResponse": {
        "$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_BodyPath_InitiateRegistrationData___"
      },
...
      "SuccessResponseType_User_": {
        "description": "Wrapper de éxito JSend (aplicado por el middleware global).",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "success"
            ],
            "nullable": false
          },
          "data": {
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "status",
          "data"
        ],
        "type": "object",
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "jwt": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "jwt-optional": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "info": {
    "title": "@flaights/backend",
    "version": "1.0.0",
    "license": {
      "name": "ISC"
    },
    "contact": {}
  },
  "paths": {
    "/users/register/initiate": {
      "post": {
        "operationId": "InitiateRegistration",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseType"
                }
              }
            }
          },
          "409": {
            "description": "Email ya registrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_EmailAlreadyInUseError_"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiateRegistrationRequestValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Paso 1 del registro: Inicia el registro enviando un código de verificación al email.",
        "tags": [
          "Users"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateRegistrationData"
              }
            }
          }
        }
      }
    },
    "/users/register/complete": {
      "post": {
        "operationId": "CompleteRegistration",
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseType_User_"
                }
              }
            }
          },
          "400": {
            "description": "Código inválido o expirado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_EmailVerificationCodeInvalidOrExpiredError_"
                }
              }
            }
          },
          "409": {
            "description": "Email ya registrado",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/FailResponseFromError_EmailAlreadyInUseError_"
                    },
                    {
                      "$ref": "#/components/schemas/FailResponseFromError_UsernameAlreadyInUseError_"
                    }
                  ]
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteRegistrationRequestValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Paso 2 del registro: Completa el registro verificando el código y creando el usuario.",
        "tags": [
          "Users"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteRegistrationData"
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "operationId": "GetSelfUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseType_GetUserResponseData_"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          }
        },
        "description": "Obtiene los datos del usuario autenticado.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": []
      },
      "patch": {
        "operationId": "UpdateUser",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseType_UpdateUserResponseData_"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_UserNotFoundError_"
                }
              }
            }
          },
          "409": {
            "description": "El nombre de usuario ya está en uso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_UsernameAlreadyInUseError_"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UpdateUserValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Actualiza los datos del usuario autenticado (sin incluir email).",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserData"
              }
            }
          }
        }
      }
    },
    "/users/me/change-email/initiate": {
      "post": {
        "operationId": "InitiateEmailChange",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseType"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          },
          "409": {
            "description": "El nuevo email ya está en uso",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_EmailAlreadyInUseError_"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InitiateEmailChangeRequestValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Inicia el proceso de cambio de email enviando códigos al email antiguo y al nuevo.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InitiateEmailChangeData"
              }
            }
          }
        }
      }
    },
    "/users/me/change-email/complete": {
      "post": {
        "operationId": "CompleteEmailChange",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseType_User_"
                }
              }
            }
          },
          "400": {
            "description": "Códigos inválidos o expirados",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_EmailVerificationCodeInvalidOrExpiredError_"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompleteEmailChangeRequestValidationFailResponse"
                }
              }
            }
          }
        },
        "description": "Completa el cambio de email verificando ambos códigos.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteEmailChangeData"
              }
            }
          }
        }
      }
    },
    "/users/me/change-email/cancel": {
      "post": {
        "operationId": "CancelEmailChange",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseType"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_UserNotFoundError_"
                }
              }
            }
          }
        },
        "description": "Cancela el cambio de email pendiente.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": []
      }
    },
    "/users/search": {
      "get": {
        "operationId": "SearchUsers",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseType_PublicUser-Array_"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_QueryPath__q-string____"
                }
              }
            }
          }
        },
        "description": "Busca usuarios.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/users/{id}": {
      "get": {
        "operationId": "GetUserById",
        "responses": {
          "200": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponseType_GetUserByIdResponseData_"
                }
              }
            }
          },
          "401": {
            "description": "No autenticado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthFailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Usuario no encontrado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FailResponseFromError_UserNotFoundError_"
                }
              }
            }
          },
          "422": {
            "description": "Error de validación",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestValidationFailResponse_ValidationDetails_PathPath__id-string____"
                }
              }
            }
          }
        },
        "description": "Obtiene los datos de un usuario por su ID.",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "jwt": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/admin/users/{id}": {
      "delete": {
        "operationId": "DeleteUser",
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "description": "Eliminar un usuario permanentemente (Solo Admin).",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "jwt": [
              "admin",
              "superadmin"
            ]
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "servers": [
    {
      "url": "/"
    }
  ]
}